拆分添加下载连接
