+++ src/dotnetCampus.FileDownloader.Tool/dotnetCampus.FileDownloader.Tool.csproj
@@ -3,6 +3,10 @@
     <PropertyGroup>
         <OutputType>Exe</OutputType>
         <TargetFramework>netcoreapp3.1</TargetFramework>
+
+        <PackAsTool>true</PackAsTool>
+
+        <ToolCommandName>FileDownloader</ToolCommandName>
     </PropertyGroup>
 
     <ItemGroup>
