+++ src/dotnetCampus.FileDownloader.WPF/ObservableCollectionExtension.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
+
+namespace dotnetCampus.FileDownloader.WPF
+{
+    public static class ObservableCollectionExtension
+    {
+        public static void AddRange<T>(this ObservableCollection<T> observableCollection, IEnumerable<T> list)
+        {
+            foreach (var temp in list)
+            {
+                observableCollection.Add(temp);
+            }
+        }
+    }
+}
